♻️ [REFACTOR] 배틀 모드 구조 변경
